Commit the initial spike code
I have scaffolded the repository and implemented the basics needed to display the canvas and react to the code input in the text area. I have not found the final solution for the point to point arc yet, so I have hacked around it by using a quadratic curve for now. I have bookmarked a few links to look into later.