Add support for showing URL-based references
I've put some effort into making the references only download once, but there are still unanswered questions, like what should happen if multiple references are specified. I think there is no harm in showing them all in the document order? For now I will leave it like this. Another thing is the need for supporting local references, but that's a bit more tricky as I will need to update the source code based on the file picker name and I will also need to open the picker for the right reference being edited, so I am leaving that as a plan level for now.