Prevent negative zoom which flips signs
It was possible to go below zero and the negative zoom would flip signs and as a result flip the whole sketch. I have capped it at a zero because I do not feel comfortable hardcoding in any particular minimal threshold, so I won't take away the user's ability to shrink everything to a single point.