Fix the order of zoomDiv and coordsDiv to match DOM
At some point I swapped the two as the coordsDiv has a dynamic size and changes on every mouse move and the zoomDiv is more static, so this order makes them less distruptive, but I forgot to change the order in the code. The order of their event handlers is okay now.