Implement loading a saved file
I have decided to use a file input element that stays in the DOM as opposed to creating one on the flies so that the file open dialog remembers the last selected directory.

Opening the file will rewrite the draft saved in the local storage and it won't ask whether to really open if the draft is currently unsaved; I don't track that information at all. I am okay with both of these behaviors for now, but may improve them in the future.