Implement opening local reference images
URL-based references are already supported and the user can go in and type the command including the URL and get the reference showing.

This change introduces a new menu button, Reference, which presents a file picker and the selected image gets cached and displayed as a reference in the same way.

The corresponding reference command gets placed at the top of the source code and uses the image name to find the image in the cache.

For now, I don't handle the user changing the command's reference name in any way other than the default download failure callback. I have left some tasks to look at improving this UX in the future.

I have decided against opening the reference modal on the name change and replacing the unknown name with the selected file. It seems like weird UX not worth the effort of associating the new file with the changed reference command (there can be multiple).