Implement rendering out to an image and plan improvements
This feature allows the user to render the sketch to an image file. It won't work in case URL references are used, it is a browser limitation due to CORS. I have noted a few ideas I have around how to work around that in a way that is as smooth as possible for the user.