Handle empty and commented out lines separately
These lines do not have any impact on the sketch's rendering, but they still need an associated hint so that the hints do not go offset from the lines they belong to. We can do that in one step before the switch for the real commands. Doing this way has an added benefit of the commented out line not needing to start with "// " (with space) but can start with just "//" (no space).