Add instructions for running off file with ESM and make main script a module
I plan on breaking down the script files into multiple ones each hosting its own function or bit of functionality. In order to do that, I will need ESM as I do not want to use multiple script files. In order to use ESM in the browser over the file protocol, Firefox needs this hack to stop enforcing CORS. This is a dangerous setting to turn off when planning on using Firefox with untrusted local files, but I never do that.