Pull out individual command handlers to their own files
I've replaced the big switch with a map from command name or alias to the implementation function and gave the handlers all the same interface, which is expressive enough that it can be used to implement all the handlers I have so far. I expect I will tweak this interface in the future as I add new commands and their handlers.