Add a few more notes and thoughts
I am finally starting to see why all of the arc APIs seem so needlessly complex... :) I used to think I would be fine with the two points and could derive the rest from context. Which might still be true, I could track the knowledge of the shape coming in and out of the arc and calculate the slopes that would become the inputs to the arc, but it will probably make more sense to just add support for all of Bezier, quadratic and ellipse and use whichever fits the underlying reference the best. To be able to do this efficiently, I am starting to lean towards introducing handles to tweak the curves by. This would need 1:1 sync between the preview and the mouse actions and the source code, which is what I wanted to avoid, but maybe it makes the most sense?