Add more thoughts on possible solutions to smooth corners
I also came up with the idea that maybe I should split the tracing process and the process of finding corners and spikes that are too sharp and fitting these smoothing arcs in there. It can be imagined like having a shape and tracing it with a very thick brush from within - it will not fit into all corners and spikes, so the outline of the shape the thick brush would draw is the outer shape I need to derive from the original traced shape.